Added documentation to deserializer classes, removed unnecessary questionsTags junction, renamed tag package
