+++ src/main/java/com/revaturelabs/ask/question/QuestionController.java
@@ -19,7 +19,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
-import com.revaturelabs.ask.tags.TagService;
+import com.revaturelabs.ask.tag.TagService;
 import com.revaturelabs.ask.user.UserNotFoundException;
 
 /**
