--- src/main/java/com/revaturelabs/ask/questionTagsJunction/QuestionTagsJunctionService.java
@@ -1,7 +1,0 @@
-package com.revaturelabs.ask.questionTagsJunction;
-
-public interface QuestionTagsJunctionService {
-  
-  void create(int questionId, String [] tagName);
-
-}
