+++ src/main/java/com/revaturelabs/ask/tag/TagController.java
@@ -1,4 +1,4 @@
-package com.revaturelabs.ask.tags;
+package com.revaturelabs.ask.tag;
 
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
